refactor: Move mapper to adapters/mysql
