+++ backend_rust/src/adapters/repository/db_recipe_repository.rs
@@ -2,9 +2,9 @@ use crate::application::port::repository::RecipeRepository;
 use sea_orm::DatabaseConnection;
 use std::sync::Arc;
 
-use crate::adapters::recipe_mapper::RecipeMapper;
 use crate::adapters::repository::mysql::entity as db_entity;
-use crate::adapters::step_mapper::StepMapper;
+use crate::adapters::repository::mysql::recipe_mapper::MysqlRecipeMapper;
+use crate::adapters::repository::mysql::step_mapper::MysqlStepMapper;
 use crate::application::mapper::db_mapper::DbMapper;
 use crate::domain::Recipe;
 use crate::domain::Step;
@@ -25,7 +25,10 @@ impl RecipeRepository for DbRecipeRepository {
             .await
             .unwrap();
         let (recipe_model, step_models) = recipe_with_steps.pop().unwrap();
-        let steps: Vec<Step> = step_models.into_iter().map(StepMapper::to_entity).collect();
+        let steps: Vec<Step> = step_models
+            .into_iter()
+            .map(MysqlStepMapper::to_entity)
+            .collect();
 
         Ok(Recipe {
             id: recipe_id,
@@ -49,8 +52,8 @@ impl RecipeRepository for DbRecipeRepository {
 
         let mut recipes: Vec<Recipe> = Vec::new();
         for (recipe, steps) in recipe_models {
-            let steps: Vec<Step> = steps.into_iter().map(StepMapper::to_entity).collect();
-            let mut recipe = RecipeMapper::to_entity(recipe);
+            let steps: Vec<Step> = steps.into_iter().map(MysqlStepMapper::to_entity).collect();
+            let mut recipe = MysqlRecipeMapper::to_entity(recipe);
             recipe.steps = steps;
             recipes.push(recipe);
         }
@@ -63,7 +66,7 @@ impl RecipeRepository for DbRecipeRepository {
             .await?;
         let recipes = recipe_models
             .into_iter()
-            .map(RecipeMapper::to_entity)
+            .map(MysqlRecipeMapper::to_entity)
             .collect();
         Ok(recipes)
     }
