+++ backend_rust/src/adapters/repository/mysql.rs
@@ -1,2 +1,4 @@
 pub mod entity;
+pub mod recipe_mapper;
 pub mod resource_mapper;
+pub mod step_mapper;
