+++ backend_rust/src/adapters/step_mapper.rs
@@ -1,8 +1,5 @@
-use crate::adapters::repository::mysql::entity::steps::Model as StepModel;
 use crate::application::mapper::api_mapper::ApiMapper;
-use crate::application::mapper::db_mapper::DbMapper;
 use crate::domain::Step;
-use crate::domain::entity::resource::ResourceId;
 use crate::presentation::graphql::object::Step as StepObject;
 pub struct StepMapper {}
 
@@ -14,22 +11,6 @@ impl ApiMapper<Step, StepObject> for StepMapper {
             resource_id: entity.resource_id.0 as u64,
             order_number: entity.order,
             duration: entity.duration,
-        }
-    }
-}
-
-impl DbMapper<Step, StepModel> for StepMapper {
-    fn to_db(_entity: Step) -> StepModel {
-        unimplemented!()
-    }
-
-    fn to_entity(model: StepModel) -> Step {
-        Step {
-            id: model.id,
-            description: model.description,
-            order: model.order_number,
-            duration: model.duration,
-            resource_id: ResourceId(model.resource_id as i32),
         }
     }
 }
